win-tray: move cli back to main
parts needed in gui were duped, so cli doesn't need to be a lib anymore
